feat: optimize common component
