+++ packages/next-common/components/collectives/members/table.js
@@ -1,7 +1,6 @@
 import FellowshipRank from "next-common/components/fellowship/rank";
 import ValueDisplay from "next-common/components/valueDisplay";
 import { toPrecision } from "next-common/utils";
-import Period from "next-common/components/fellowship/params/period";
 import { useNavCollapsed } from "next-common/context/nav";
 import AddressUser from "next-common/components/user/addressUser";
 import DataList from "next-common/components/dataList";
@@ -9,7 +8,7 @@ import { isNil } from "lodash-es";
 import collectivesMemberColumns from "next-common/components/collectives/members/columns";
 import rankToIndex from "next-common/utils/fellowship/rankToIndex";
 import { getRankSalary } from "next-common/utils/fellowship/getRankSalary";
-import MemberPeriodProgress from "next-common/components/collectives/members/periodProgress.jsx";
+import MemberPeriodWithProgress from "next-common/components/collectives/members/periodWithProgress.jsx";
 
 function AddressCol({ address }) {
   const [navCollapsed] = useNavCollapsed();
@@ -31,6 +30,9 @@ export default function CollectivesMemberTable({
   } = params ?? {};
   const { symbol, decimals } = salaryAsset;
 
+  const isLoading =
+    isNil(members) || isNil(params) || isNil(salaryAsset) || isNil(coreMembers);
+
   const rows = (members || []).map(({ address, rank }, idx) => {
     return [
       <FellowshipRank key={`rank-row-${idx}`} rank={rank} />,
@@ -45,37 +47,31 @@ export default function CollectivesMemberTable({
         value={toPrecision(getRankSalary(passiveSalary, rank), decimals)}
         symbol={symbol}
       />,
-      <div className="max-sm:text-right">
-        <Period
-          key={`demotion-period-${idx}`}
-          blocks={demotionPeriod[rankToIndex(rank)] || offboardTimeout}
-        />
-        <MemberPeriodProgress
+      demotionPeriod[rank] > 0 && (
+        <MemberPeriodWithProgress
+          keyPrefix={`demotion-period-${idx}`}
+          periodKey={rankToIndex(rank)}
           members={members}
           address={address}
           params={params}
           rank={rank}
           coreMembers={coreMembers}
           type="demotion"
+          blocks={demotionPeriod[rankToIndex(rank)] || offboardTimeout}
         />
-      </div>,
-      rank === 0 && minPromotionPeriod[rank] <= 0 ? (
-        ""
-      ) : (
-        <div className="max-sm:text-right">
-          <Period
-            key={`min-promotion-period-${idx}`}
-            blocks={minPromotionPeriod[rank] || 0}
-          />
-          <MemberPeriodProgress
-            members={members}
-            address={address}
-            params={params}
-            rank={rank}
-            coreMembers={coreMembers}
-            type="promotion"
-          />
-        </div>
+      ),
+      minPromotionPeriod[rank] > 0 && (
+        <MemberPeriodWithProgress
+          keyPrefix={`promotion-period-${idx}`}
+          periodKey={rank}
+          members={members}
+          address={address}
+          params={params}
+          rank={rank}
+          coreMembers={coreMembers}
+          type="promotion"
+          blocks={minPromotionPeriod[rank] || 0}
+        />
       ),
     ];
   });
@@ -86,7 +82,7 @@ export default function CollectivesMemberTable({
       columns={collectivesMemberColumns}
       noDataText="No Members"
       rows={rows}
-      loading={isNil(members)}
+      loading={isLoading}
     />
   );
 }
